Cleanup Startup class to make it easier to understand
